@gui @text @gfix @test @change @int @float @refactor
[src/fsix.hxx]
> replaced all floats with ints
[src/gfix.hxx]
> replaced all floats with ints
> added quad_t and text_t structures
> implemented stroke text rendering
> declared drawing utility functions
> > these are classified as "global" and "screen" scoped functions
> > > "global" means in-world objects
> > > "screen" means on-screen objects
> > for now there are 3
> > > draw_global_quad
> > > draw_screen_quad
> > > draw_screen_text
> declared external draw_loop callback
> changed scaling behaviour
> heavily relying on integer values
> using only rgb, no alpha anymore
> made the screen center as a coordinate system origin
> added enum to define coordinates origin
> > used by screen drawing functions to draw from a keypoint
[src/gfix.cxx]
> moved in all graphics related logic
> added drawing utility functions
[src/head.hxx]
> added more standard headers
> > include <algorithm>
> > > for std::clamp
> > include <cstdint>
> > > just in case
[src/iput.cxx]
> changed key_line content for the "invalid keybind" case
> just do "key_line_insert" for proc_key_board_special
> > which means that all special chars are treated as normal chars
[src/main.hxx]
> replaced all floats with ints
> removed unused in this header headers
> removed unused function declarations
[src/main.cxx]
> replaced all floats with ints
> added required headers
> changed timer logic
> > update fps once per second
> removed rendering logic
> removed useless output
> simplified callbacks
> added glutSetKeyRepeat
[Makefile]
> increased the version to 0xa0a1a0

+++ src/fsix.hxx
@@ -10,10 +10,10 @@ _NAMESPACE_ENTER
 
 typedef struct com_fsix_body_t
 {
-    float rotation = 0.0;
-    float rotation_velocity = 0.0;
-    struct { float x = 0.0, y = 0.0; } position;
-    struct { float x = 0.0, y = 0.0; } position_velocity;
+    int rotation = 0;
+    int rotation_velocity = 0;
+    struct { int x = 0, y = 0; } position;
+    struct { int x = 0, y = 0; } position_velocity;
 } com_fsix_body_t;
 
 typedef struct fsix_t
